refactor: break down monolithic API into focused modules
- Split main.py (825 lines) into focused modules:
  * routes/sessions.py - Session management endpoints
  * routes/tutoring.py - Tutoring interaction endpoints
  * routes/visualization.py - Visualization endpoints
  * routes/health.py - Health check endpoints
- Extract Pydantic models into separate files:
  * models/requests.py - Request models
  * models/responses.py - Response models
- Create middleware package for cross-cutting concerns:
  * middleware/cors.py - CORS configuration
  * middleware/error_handling.py - Exception handlers
  * middleware/logging.py - Request logging
- Implement dependencies.py for proper dependency injection
- Streamline main.py to only handle app initialization and routing
- Maintain backward compatibility with all existing endpoints
