+++ backend/src/api/middleware/cors.py
@@ -1,0 +1,17 @@
+"""
+CORS middleware configuration.
+"""
+
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+
+
+def setup_cors(app: FastAPI) -> None:
+    """Setup CORS middleware for the FastAPI application."""
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=["*"],  # In production, specify exact origins
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
