+++ backend/src/api/routes/__init__.py
@@ -1,0 +1,16 @@
+"""
+API routes package for organizing endpoint handlers.
+"""
+
+# Import route modules
+from . import sessions
+from . import tutoring  
+from . import visualization
+from . import health
+
+__all__ = [
+    "sessions",
+    "tutoring",
+    "visualization", 
+    "health"
+]
